Split users.go into dedicated parts
